+++ app/models/descriptor.rb
@@ -13,6 +13,7 @@ class Descriptor < ApplicationRecord
   include Shared::Identifiers
   include Shared::Tags
   include Shared::Notes
+  include Shared::Depictions
   include Shared::DataAttributes
   include Shared::AlternateValues
   include Shared::Confidences
@@ -24,7 +25,7 @@ class Descriptor < ApplicationRecord
 
   acts_as_list scope: [:project_id]
 
-  ALTERNATE_VALUES_FOR = [:name, :short_name].freeze
+  ALTERNATE_VALUES_FOR = [:name, :short_name, :description_name, :key_name].freeze
 
   validates_presence_of :name, :type
   validate :type_is_subclassed
@@ -39,8 +40,35 @@ class Descriptor < ApplicationRecord
 
   soft_validate(:sv_short_name_is_short)
 
+  scope :not_weight_zero, -> {where('NOT "descriptors"."weight" = 0 OR "descriptors"."weight" IS NULL') }
+
   def self.human_name
     self.name.demodulize.humanize
+  end
+
+  # @return [String] name of the descriptor in a particular language
+  # @params target [Symbol] one of :key, :description, nil
+  # TODO: This should be a helper method, not a model method
+  def target_name(target, language_id)
+    n = self.name
+    a = nil
+    case target
+    when :key
+      n = self.key_name.nil? ? n : self.key_name
+    when :description
+      n = self.description_name.nil? ? n : self.description_name
+    end
+    unless language_id.nil?
+      case target
+      when :key
+        a = AlternateValue::Translation.where(alternate_value_object: self, language_id: language_id, alternate_value_object_attribute: 'key_name').first
+      when :description
+        a = AlternateValue::Translation.where(alternate_value_object: self, language_id: language_id, alternate_value_object_attribute: 'description_name').first
+      end
+      a = AlternateValue::Translation.where(alternate_value_object: self, language_id: language_id, alternate_value_object_attribute: 'name').first if a.nil?
+      n = a.value unless a.nil?
+    end
+    return n
   end
 
   def qualitative?