+++ app/models/observation.rb
@@ -23,7 +23,6 @@ class Observation < ApplicationRecord
   belongs_to :collection_object, inverse_of: :observations
 
   before_validation :convert_observation_object_global_id
-
   before_validation :set_type_from_descriptor
 
   validates_presence_of :descriptor_id, :type