+++ CHANGELOG.md
@@ -8,8 +8,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ## [unreleased]
 
 ### Added
-
-- Added ancestors facet on filter nomenclature [#1791]
+- New Interactive Key task [#1810] 
+- New model for matrix based interactive keys which produce JSON for the Interactive Key task [#1810]
+- `weight` field to descriptor
+- Ancestors facet on filter nomenclature [#1791]
 - TW_DISABLE_DB_BACKUP_AT_DEPLOY_TIME env var to disable built-in backup functionality at deploy/database-update time.
 - Display coordinate type specimens [#1811]
 - Changed background color header for invalid names on Browse OTU
@@ -36,6 +38,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#1791]: https://github.com/SpeciesFileGroup/taxonworks/issues/1791 
 [#1787]: https://github.com/SpeciesFileGroup/taxonworks/issues/1787
 [#1798]: https://github.com/SpeciesFileGroup/taxonworks/issues/1798
+[#1810]: https://github.com/SpeciesFileGroup/taxonworks/pull/1810
 [#1811]: https://github.com/SpeciesFileGroup/taxonworks/issues/1811